Avatar get from server, rewrite code.
