+++ .gitignore
@@ -1,3 +1,5 @@
 /node_modules
 /dist
 /.vscode
+/.DS_Store
+/.idea
