+++ README.md
@@ -1,1 +1,7 @@
-# Places
+#**Проектная работа №11**
+
+[Ссылка на мою работу](https://ppridanov.github.io/Places "Проектная работа №11")
+
+My third github project
+
+version 1.0.2
