--- blocks/place-card/__image/place-card__image.css
@@ -1,18 +1,0 @@
-.place-card__image {
-  height: 282px;
-  background-size: cover;
-  background-position: center;
-  border-top-left-radius: 10px;
-  border-top-right-radius: 10px;
-  position: relative;
-}
-
-.place-card__image::after {
-  content: '';
-  display: block;
-  width: 282px;
-  height: 282px;
-  background-color: rgba(0, 0, 0, .3);
-  border-top-left-radius: 10px;
-  border-top-right-radius: 10px;
-}
