+++ blocks/user-info/user-info.css
@@ -1,9 +1,9 @@
-@import url(./__photo/user-info__photo.css);
-@import url(./__data/user-info__data.css);
-@import url(./__name/user-info__name.css);
-@import url(./__job/user-info__job.css);
 @import url(./__add-button/user-info__add-button.css);
+@import url(./__data/user-info__data.css);
 @import url(./__edit-button/user-info__edit-button.css);
+@import url(./__job/user-info__job.css);
+@import url(./__name/user-info__name.css);
+@import url(./__photo/user-info__photo.css);
 
 .user-info {
   display: flex;
