+++ src/modules/card.js
@@ -1,15 +1,16 @@
-import {placeList} from "./script.js";
-import {connection} from "../modules/api.js";
+import {placeList} from "./listcard.js";
+import {connection} from "./api.js";
 export default class Card {
   constructor(name, link, id, ownerId, likesArray) {
+    if (likesArray == undefined) likesArray = false;
     this.cardContainer = this.create(name, link, id, likesArray.length);
     const likeCounter = this.cardContainer.querySelector('.place-card__like-counter');
     this.cardContainer.querySelector('.place-card__like-icon').addEventListener('click', (event) => {
-      this.like(event, id, likeCounter, likesArray.length)
+      this.like(event, id, likeCounter, likesArray.length);
     });
     this.cardContainer.querySelector('.place-card__delete-icon').addEventListener('click', (event) => {
       if (ownerId == myOwnerId) {
-        this.remove(event)
+        this.remove(event);
       }
     })
     this.cardContainer.querySelector('.place-card__image').addEventListener('click', this.image);
@@ -19,23 +20,26 @@ export default class Card {
       removeButton.classList.remove('place-card__delete-icon');
       removeButton.classList.add('place-card__delete-icon_not-my');
     }
-    this.checkLike(likesArray, myOwnerId);
+    if (likesArray) {
+      this.checkLike(likesArray, myOwnerId);
+    }
   }
   create(nameValue, linkValue, idValue, likesValue) {
+    if (likesValue == undefined) likesValue = 0;
     const cardContainer = document.createElement('div');
     cardContainer.classList.add('place-card');
     const templateCards = `
       <div class='place-card__container-image' _id='${idValue}'>
-      <div class='place-card__image' style='background-image: url(${linkValue})'></div>
+        <div class='place-card__image' style='background-image: url(${linkValue})'></div>
         <button class='place-card__delete-icon'></button>
       </div>
       <div class='place-card__description'>
-        <h3 class='place-card__name'>${nameValue}</h3>
-        <div class='place-card__like-container'>
-          <button class='place-card__like-icon'></button>
-          <p class='place-card__like-counter'>${likesValue}</p>
-        </div>
-        `;
+      <h3 class='place-card__name'>${nameValue}</h3>
+      <div class='place-card__like-container'>
+        <button class='place-card__like-icon'></button>
+        <p class='place-card__like-counter'>${likesValue}</p>
+      </div>
+      `;
     cardContainer.innerHTML = templateCards;
     return cardContainer; 
   }
@@ -56,15 +60,15 @@ export default class Card {
     }
   }
   checkLike(likesArray, myId) {
-    likesArray.forEach((item) => {
-      if (item._id === myId) {
-        const likeIcon= this.cardContainer.querySelector('.place-card__like-icon');
-        likeIcon.classList.add('place-card__like-icon_liked');
-      }
-    })
+      likesArray.forEach((item) => {
+        if (item._id === myId) {
+          const likeIcon= this.cardContainer.querySelector('.place-card__like-icon');
+          likeIcon.classList.add('place-card__like-icon_liked');
+        }
+      })
   }
   remove(event) {
-    let confirmDelete = confirm('Вы уверены что хотите удалить файл?')
+    let confirmDelete = confirm('Вы уверены что хотите удалить файл?');
     if (confirmDelete) {
       let card = event.target.closest('.place-card__container-image');
       connection.deleteCard(card.getAttribute('_id'));
@@ -81,6 +85,6 @@ export default class Card {
     const popupCloseButton = document.createElement('button');
     popupCloseButton.classList.add('popup__close');
     popupImageContainer.appendChild(openImage, popupCloseButton);
-    bigImage.classList.add('popup_is-opened')
+    bigImage.classList.add('popup_is-opened');
   }
 }
