+++ src/modules/validate.js
@@ -1,10 +1,17 @@
-import {authorInput, jobInput, nameInput, linkInput, submitButton, validator, renderLoading, userInfo, popup, listCard} from "./script.js";
-import {connection} from "../modules/api.js";
+import {connection} from "./api.js";
+import {listCard} from './listcard.js';
+import {popup} from "./popup.js";
+const authorInput = document.querySelector('.popup__input_type_author');
+const jobInput = document.querySelector('.popup__input_type_job');
+const nameInput = document.querySelector('.popup__input_type_name');
+const linkInput = document.querySelector('.popup__input_type_link');
+const submitButton = document.querySelectorAll('.popup__button');
+
+
 export default class Validate {
   constructor() {
     let editForm = document.forms.edit;
     let newForm = document.forms.new;  
-
     authorInput.addEventListener('input', (event) => {
       this.handleValidate(event);
       this.checkEditForm(editForm, submitButton);
@@ -46,11 +53,11 @@ export default class Validate {
       } 
       if (element.name != submitButton[0].name && element.name === nameInput.name) {
         if (element.value.length <= 1 || element.value.length >= 30) {
-        errorElement.textContent = errorMessages.little;
-        this.activateError(element);
-        return false; 
+          errorElement.textContent = errorMessages.little;
+          this.activateError(element);
+          return false; 
+        }
       }
-    }
     this.enableButton();
     return true;
   }
@@ -86,12 +93,11 @@ export default class Validate {
   }
   checkPlaceForm(form) {
     let isValidForm = true; 
-    console.log(form)
     const inputs1 = Array.from(form.elements);
     inputs1.forEach(elem => {
       if (!this.validate(elem)) {
-          isValidForm = false;
-        }
+        isValidForm = false;
+      }
     });
     if (isValidForm) {
       this.enableButton();
@@ -136,9 +142,22 @@ export default class Validate {
         connection.editProfile(editForm.elements.author.value, editForm.elements.job.value)
           .then((result) => {
               popup.close(event)
+              const nameValue = document.querySelector('.user-info__name');
+              const jobValue = document.querySelector('.user-info__job');
+              nameValue.textContent = result.name;
+              jobValue.textContent = result.about;
           })
-        userInfo.render(editForm.elements.author.value, editForm.elements.job.value)         
       };
     }
   }
 }
+const validator = new Validate;
+function renderLoading(isLoading) {
+  const popupButton = document.querySelector('.popup__button');
+  if (isLoading) {
+    popupButton.textContent = 'Загрузка...';
+  } else {
+    popupButton.textContent = 'Сохранить';
+  }
+}
+export {validator, renderLoading};
