+++ src/script.js
@@ -1,0 +1,8 @@
+import "./modules/card.js"
+import "./modules/listcard.js";
+import "./modules/userinfo.js";
+import "./modules/popup.js";
+import "./modules/validate.js";
+import "./modules/api.js"
+import './style.css'
+
