+++ src/style.css
@@ -1,1 +1,16 @@
-@import url('../pages/index.css');
+@import url(./vendor/normalize.css);
+@import url(./vendor/fonts.css);
+@import url(../blocks/root/root.css);
+@import url(../blocks/header/header.css);
+@import url(../blocks/logo/logo.css);
+@import url(../blocks/profile/profile.css);
+@import url(../blocks/user-info/user-info.css);
+@import url(../blocks/button/button.css);
+@import url(../blocks/places-list/places-list.css);
+@import url(../blocks/place-card/place-card.css);
+@import url(../blocks/popup/popup.css);
+@import url(../blocks/error/error.css);
+
+body {
+  background-color: #000;
+}
